Changed the name to a more suitable library name  Changes to be committed: 	modified:   README.md 	modified:   core/pom.xml 	modified:   pom.xml 	renamed:    listenhere4sqs/pom.xml -> qcat4sqs/pom.xml 	renamed:    listenhere4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4Sqs.java -> qcat4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4Sqs.java 	renamed:    listenhere4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4SqsHelper.java -> qcat4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4SqsHelper.java 	renamed:    listenhere4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4SqsWorker.java -> qcat4sqs/src/main/java/net/subnoize/listenhere/sqs/ListenHere4SqsWorker.java 	renamed:    listenhere4sqs/src/main/java/net/subnoize/listenhere/sqs/SqsExecutionTemplate.java -> qcat4sqs/src/main/java/net/subnoize/listenhere/sqs/SqsExecutionTemplate.java 	renamed:    listenhere4sqs/src/main/java/net/subnoize/listenhere/sqs/SqsMessageAttributes.java -> qcat4sqs/src/main/java/net/subnoize/listenhere/sqs/SqsMessageAttributes.java 	renamed:    listenhere4sqs/src/main/resources/META-INF/spring.factories -> qcat4sqs/src/main/resources/META-INF/spring.factories 	renamed:    listenhere4sqs/src/test/resources/.gitkeep -> qcat4sqs/src/test/resources/.gitkeep
