+++ README.md
@@ -1,2 +1,2 @@
-# ListenHere
+# QCat &trade;
 Spring friendly messaging system integration framework.
