+++ core/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>listenhere-core</artifactId>
+	<artifactId>qcat-core</artifactId>
 	<packaging>jar</packaging>
 	<name>Core</name>
 	<description>The core library of components and runtime configuration</description>
@@ -9,8 +9,8 @@
 	<!--The identifier of the parent POM -->
 	<parent>
 		<groupId>net.subnoize</groupId>
-		<artifactId>listenhere</artifactId>
-		<version>0.0.3-RELEASE</version>
+		<artifactId>qcat</artifactId>
+		<version>0.0.4-SNAPSHOT</version>
 	</parent>
 
 	<properties>
